Вынес шаблонов  и их наследование. Создание Блюпринтов
