+++ webapp/db.py
@@ -1,8 +1,8 @@
+
 from flask_sqlalchemy import SQLAlchemy
-
-
 
 db=SQLAlchemy()
 
 class Base(db.Model): 
     __abstract__ = True #  класс абстрактный
+
