--- webapp/templates/chekmail.html
@@ -1,38 +1,0 @@
-<!DOCTYPE html>
-<html lang="ru">
-
-<head>
-  <meta charset="UTF-8">
-  <meta http-equiv="X_UA-Comatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet"
-    integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
-  <title>{{ page_title }}</title>
-  <link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='style.css') }}">
-</head>
-
-
-<body>
-  <section id="chekmail">
-    <div class="card-body">
-      <div class="row justify-content-center align-items-center vh-100">
-        <div class="col-md-6 col-lg-5">
-          <form class="border border-collapse p-4 rounded">
-            <div class="mb-3">
-              <label for="InputEmail2" class="form-label">Введите Почту для восстановления доступа:</label>
-              <input type="email" class="form-control" id="InputEmail2" aria-describedby="emailHelp">
-              <div id="emailHelp" class="form-text">{{ form.chek_email.label }}</div>
-              <button type="button" class="btn btn-secondary  w-50">{{ form.start_chek_and_send.label }}</button>
-            </div>
-          </form>
-        </div>
-      </div>
-    </div>
-  </section>
-  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
-    integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
-    crossorigin="anonymous"></script>
-  <script src="/static/script.js"></script>
-</body>
-
-</html>
