+++ webapp/templates/login/chekmail.html
@@ -1,0 +1,21 @@
+{% extends "base.html" %}
+
+{% block content %}
+
+  <section id="chekmail">
+    <div class="card-body">
+      <div class="row justify-content-center align-items-center vh-100">
+        <div class="col-md-6 col-lg-5">
+          <form class="border border-collapse p-4 rounded">
+            <div class="mb-3">
+              <label for="InputEmail2" class="form-label">Введите Почту для восстановления доступа:</label>
+              <input type="email" class="form-control" id="InputEmail2" aria-describedby="emailHelp">
+              <div id="emailHelp" class="form-text">{{ form.chek_email.label }}</div>
+              <button type="button" class="btn btn-secondary  w-50">{{ form.start_chek_and_send.label }}</button>
+            </div>
+          </form>
+        </div>
+      </div>
+    </div>
+  </section>
+  {% endblock %}
