+++ webapp/templates/login/login.html
@@ -1,21 +1,7 @@
-<!DOCTYPE html>
-<html lang="ru">
+{% extends "base.html" %}
 
-<head>
-  <meta charset="UTF-8">
-  <meta http-equiv="X_UA-Comatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <!-- <title>Login</title> -->
-  <!-- <link rel="stylesheet" href="https://cdn.lineicons.com/5.0/lineicons.css" /> -->
-  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet"
-    integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
-  <title>{{ page_title }}</title>
-  <link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='style.css') }}">
-</head>
-
-
-<body>
-  <section id="login">
+{% block content %}
+<section id="login">
     <div class="container-fluid">
       <div class="row justify-content-center align-items-center vh-100">
         <div class="col-md-6 col-lg-5">
@@ -38,7 +24,7 @@
                     </div>
                     {% endif %}
                 {% endwith %}
-                  <form action="{{ url_for('process_login') }}" method="post" class="row g-3">
+                  <form action="{{ url_for('user.process_login') }}" method="post" class="row g-3">
                     {{ login_form.hidden_tag() }}
                     <div class="col-md-12">
                       <label for="email_name_chek1" class="form-label">{{ login_form.username.label }}
@@ -58,15 +44,15 @@
                     </div>
                     <div class="d-flex justify-content-between">
                       <div class="form-check">
-                        <input type="chekbox" class="form-check-input">
-                        <label for="chekbox">{{ login_form.save_login.label }}</label>
+                        {{ login_form.save_login() }}
+                        {{ login_form.save_login.label(class="form-check-label") }}
                       </div>
                       <a href="#" class="f-passwords">
                         <u>{{ login_form.forgot_pas.label() }}</u>
                       </a>
                     </div>
                     <div class="col-md-12">
-                      <button class="btn border-0" type="submit">
+                      <button class="btn btn-primary" type="submit">
                         {{ login_form.submit_in.label }}
                       </button>
                     </div>
@@ -126,10 +112,4 @@
       </div>
     </div>
   </section>
-  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
-    integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
-    crossorigin="anonymous"></script>
-  <script src="/static/script.js"></script>
-</body>
-
-</html>
+  {% endblock %}
