+++ webapp/user/forms.py
@@ -5,18 +5,10 @@ from wtforms.validators import DataRequired, Length
 
 class LoginForm(FlaskForm):
     username = StringField("Имя пользователя или Почта:" , validators=[DataRequired()], render_kw = {"class": "form-control"})
-    password = PasswordField("Пароль:" , validators=[DataRequired(),  Length(min=2, max=20, message="Пароль должен содержать от 8 до 20 символов")], render_kw = {"class": "form-control"})
+    password = PasswordField("Пароль:" , validators=[DataRequired(),  Length(min=2, max=20, message="Пароль должен содержать от 2 до 20 символов")], render_kw = {"class": "form-control"})
     submit_in = SubmitField("Авторизация")
-    save_login = BooleanField("Запомнить меня")
+    save_login = BooleanField("Запомнить меня", default=True, render_kw={"class": "form-check-input"})
     forgot_pas = SubmitField("Пароль утерян?")
-
-class RegForm(FlaskForm):
-    label2 = StringField("Регистрация" )
-    email = EmailField("Почта:" , validators=[DataRequired()])
-    username1 = StringField("Имя пользователя:" , validators=[DataRequired()])
-    password_fir = PasswordField("Пароль: " , validators=[DataRequired(),  Length(min=8, max=20, message="Пароль должен содержать от 8 до 20 символов")])   # Ограничение на длину почты 8-20
-    password_sec = PasswordField("Подтвердите пароль: " , validators=[DataRequired(),  Length(min=8, max=20, message="Пароль должен содержать от 8 до 20 символов")])
-    submit_reg = SubmitField("Регистрация")
 
 
 class ChekMail(FlaskForm):
